refactored virt_api_pod, moved wait_for_pod_running_by_prefix to tests/utils.py to reuse it
