+++ tests/install_upgrade_operators/product_install/utils.py
@@ -3,17 +3,14 @@ import logging
 import subprocess
 
 from ocp_resources.deployment import Deployment
-from ocp_utilities.operators import TIMEOUT_5MIN
-from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
+from tests.utils import wait_for_pod_running_by_prefix
 from utilities.constants import (
     HOSTPATH_PROVISIONER,
     HOSTPATH_PROVISIONER_OPERATOR,
-    TIMEOUT_5SEC,
     TIMEOUT_15MIN,
 )
 from utilities.data_collector import write_to_file
-from utilities.infra import get_not_running_pods, get_pod_by_name_prefix
 from utilities.storage import verify_hpp_pool_health
 
 PRINT_COMMAND = '{printf "%s%s",sep,$0;sep=","}'
@@ -69,49 +66,6 @@ def get_all_resources(file_name, base_directory):
         "namespaced": namespaced_resources_dict,
         "cluster-scoped": cluster_resources_dict,
     }
-
-
-def wait_for_pod_running_by_prefix(
-    admin_client,
-    namespace_name,
-    pod_prefix,
-    expected_number_of_pods,
-    number_of_consecutive_checks=3,
-    timeout=TIMEOUT_5MIN,
-):
-    samples = TimeoutSampler(
-        wait_timeout=timeout,
-        sleep=TIMEOUT_5SEC,
-        func=get_pod_by_name_prefix,
-        dyn_client=admin_client,
-        pod_prefix=pod_prefix,
-        namespace=namespace_name,
-        get_all=True,
-    )
-    pod_names = None
-    not_running_pods = None
-    try:
-        current_check = 0
-        for sample in samples:
-            if sample:
-                not_running_pods = get_not_running_pods(pods=sample)
-                pod_names = [pod.name for pod in sample]
-                LOGGER.info(f"All {pod_prefix} pods: {pod_names}, not running: {not_running_pods}")
-                if not_running_pods:
-                    current_check = 0
-                else:
-                    if expected_number_of_pods == len(sample):
-                        current_check += 1
-                    else:
-                        current_check = 0
-            if current_check >= number_of_consecutive_checks:
-                return True
-    except TimeoutExpiredError:
-        LOGGER.error(
-            f"timeout waiting for all {pod_prefix} pods in namespace {namespace_name} to reach "
-            f"running state, out of {pod_names}, following pods are in not running state: {not_running_pods}"
-        )
-        raise
 
 
 def validate_hpp_installation(admin_client, cnv_namespace, schedulable_nodes):
