+++ tests/utils.py
@@ -18,6 +18,7 @@ from ocp_resources.kubevirt import KubeVirt
 from ocp_resources.resource import ResourceEditor
 from ocp_resources.virtual_machine import VirtualMachine
 from ocp_resources.virtual_machine_instance_migration import VirtualMachineInstanceMigration
+from ocp_utilities.operators import TIMEOUT_5MIN
 from pyhelper_utils.shell import run_ssh_commands
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
@@ -39,6 +40,8 @@ from utilities.infra import (
     get_artifactory_header,
     get_artifactory_secret,
     get_http_image_url,
+    get_not_running_pods,
+    get_pod_by_name_prefix,
 )
 from utilities.virt import (
     VirtualMachineForTests,
@@ -564,3 +567,46 @@ def create_cirros_vm(
         if wait_running:
             running_vm(vm=vm, wait_for_interfaces=False)
         yield vm
+
+
+def wait_for_pod_running_by_prefix(
+    admin_client,
+    namespace_name,
+    pod_prefix,
+    expected_number_of_pods,
+    number_of_consecutive_checks=3,
+    timeout=TIMEOUT_5MIN,
+):
+    samples = TimeoutSampler(
+        wait_timeout=timeout,
+        sleep=TIMEOUT_5SEC,
+        func=get_pod_by_name_prefix,
+        dyn_client=admin_client,
+        pod_prefix=pod_prefix,
+        namespace=namespace_name,
+        get_all=True,
+    )
+    pod_names = None
+    not_running_pods = None
+    try:
+        current_check = 0
+        for sample in samples:
+            if sample:
+                not_running_pods = get_not_running_pods(pods=sample)
+                pod_names = [pod.name for pod in sample]
+                LOGGER.info(f"All {pod_prefix} pods: {pod_names}, not running: {not_running_pods}")
+                if not_running_pods:
+                    current_check = 0
+                else:
+                    if expected_number_of_pods == len(sample):
+                        current_check += 1
+                    else:
+                        current_check = 0
+            if current_check >= number_of_consecutive_checks:
+                return True
+    except TimeoutExpiredError:
+        LOGGER.error(
+            f"timeout waiting for all {pod_prefix} pods in namespace {namespace_name} to reach "
+            f"running state, out of {pod_names}, following pods are in not running state: {not_running_pods}"
+        )
+        raise
